+++ diesel/src/expression/operators.rs
@@ -106,6 +106,21 @@ macro_rules! __diesel_operator_body {
                 Ok(())
             }
         }
+
+        impl<S, $($ty_param,)+> $crate::query_source::aliasing::FieldAliasMapper<S> for $name<$($ty_param,)+>
+        where
+            S: $crate::query_source::aliasing::AliasSource,
+            $($ty_param: $crate::query_source::aliasing::FieldAliasMapper<S>,)+
+        {
+            type Out = $name<
+                $(<$ty_param as $crate::query_source::aliasing::FieldAliasMapper<S>>::Out,)+
+            >;
+            fn map(self, alias: &$crate::query_source::Alias<S>) -> Self::Out {
+                $name {
+                    $($field_name: self.$field_name.map(alias),)+
+                }
+            }
+        }
     }
 }