+++ diesel/src/sqlite/connection/sqlite_value.rs
@@ -123,7 +123,8 @@ impl<'row, 'stmt, 'query> SqliteValue<'row, 'stmt, 'query> {
             _ => unreachable!(
                 "Sqlite's documentation state that this case ({}) is not reachable. \
                  If you ever see this error message please open an issue at \
-                 https://github.com/diesel-rs/diesel."
+                 https://github.com/diesel-rs/diesel.",
+                tpe
             ),
         }
     }