+++ diesel_cli/src/infer_schema_internals/data_structures.rs
@@ -34,7 +34,7 @@ impl fmt::Display for ColumnType {
             write!(out, "Nullable<")?;
         }
         if self.is_array {
-            write!(out, "Array<")?;
+            write!(out, "Array<Nullable<")?;
         }
         if self.is_unsigned {
             write!(out, "Unsigned<")?;
@@ -44,7 +44,7 @@ impl fmt::Display for ColumnType {
             write!(out, ">")?;
         }
         if self.is_array {
-            write!(out, ">")?;
+            write!(out, ">>")?;
         }
         if self.is_nullable {
             write!(out, ">")?;